+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Uniswap MongoDB price management 
 - Swagger documentation on /docs.
 - Added RPC Selection before minting
-- Check if gas is enough before execute transaction
+- Check if balance is enough before execute transaction
 
 ## [6.1.1] - 2022-01-27
 ### Added